Add more component inline docs
Also extracts `<PlaylistDuration>` into its own file, and renames
`<AppendToPlaylistButton>` to `<AppendAlbumTrackToPlaylistButton>`.
